+++ mimic-bukkit/src/main/kotlin/impl/quantumrpg/QuantumRpgItemsRegistry.kt
@@ -1,3 +1,22 @@
+/*
+ * This file is part of BukkitMimic.
+ * Copyright (C) 2021 Osip Fatkullin
+ * Copyright (C) 2021 EndlessCode Group and contributors
+ *
+ * BukkitMimic is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * BukkitMimic is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with BukkitMimic.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package ru.endlesscode.mimic.impl.quantumrpg
 
 import org.bukkit.inventory.ItemStack
@@ -41,7 +60,7 @@ public class QuantumRpgItemsRegistry private constructor(
             .firstOrNull()
     }
 
-    override fun getItem(itemId: String, amount: Int): ItemStack? {
+    override fun getItem(itemId: String, payload: Any?, amount: Int): ItemStack? {
         return runOnModules(itemId) { id ->
             mapNotNull { module -> module.getItemById(id) }
                 .map { it.create() }